+++ regionmask/defined_regions/from_geopandas.py
@@ -11,16 +11,17 @@ def _check_duplicates(to_check):
         return True
     else:
         if isinstance(to_check, list):
-            duplicates = set(
-                [x for x in to_check if to_check.count(x) > 1])
+            duplicates = set([x for x in to_check if to_check.count(x) > 1])
         elif isinstance(to_check, pd.core.series.Series):
             duplicates = to_check[to_check.duplicated(keep=False)]
         else:
             raise ValueError(
-                "to_check not in [list, pd.Series], found {}".format(type(to_check)))
+                "to_check not in [list, pd.Series], found {}".format(type(to_check))
+            )
         if len(duplicates) > 0:
             raise ValueError(
-                "Found duplicates {}, but should not. {}".format(duplicates, to_check))
+                "Found duplicates {}, but should not. {}".format(duplicates, to_check)
+            )
         else:
             return True
 
@@ -28,19 +29,20 @@ def _check_duplicates(to_check):
 def _construct_abbrevs(geodataframe, names):
     """Construct unique abbreviations based on geodataframe.names."""
     if names is None:
-        raise ValueError('names is None, but should be a valid column name of'
-                         'geodataframe, choose from {}'.format(geodataframe.columns))
+        raise ValueError(
+            "names is None, but should be a valid column name of"
+            "geodataframe, choose from {}".format(geodataframe.columns)
+        )
     abbrevs = []
     for item in geodataframe.T:
         name_for_abbrev = geodataframe.loc[item][names]
         # catch region with no name
         if name_for_abbrev is None:
-            name_for_abbrev = 'UND'  # for undefined
-        abbrev = "".join(word[0]
-                         for word in name_for_abbrev.split(' '))
+            name_for_abbrev = "UND"  # for undefined
+        abbrev = "".join(word[0] for word in name_for_abbrev.split(" "))
         # remove certain chars
-        for char_to_remove in ['-', ' ', ')', '(']:
-            abbrev = abbrev.replace(char_to_remove, '')
+        for char_to_remove in ["-", " ", ")", "("]:
+            abbrev = abbrev.replace(char_to_remove, "")
         # if find duplicates, add counter
         counter = 2
         if abbrev in abbrevs:
@@ -52,12 +54,7 @@ def _construct_abbrevs(geodataframe, names):
 
 
 def from_geopandas(
-    geodataframe,
-    numbers=None,
-    names=None,
-    abbrevs=None,
-    name='unnamed',
-    source=None
+    geodataframe, numbers=None, names=None, abbrevs=None, name="unnamed", source=None
 ):
     """
     create Regions from geodataframe data
@@ -90,7 +87,7 @@ def from_geopandas(
         # sort, otherwise breaks
         geodataframe = geodataframe.sort_values(numbers)
         # ensure integer
-        geodataframe[numbers] = geodataframe[numbers].astype('int')
+        geodataframe[numbers] = geodataframe[numbers].astype("int")
         numbers = _maybe_get_column(geodataframe, numbers)
     else:
         numbers = geodataframe.index.values
@@ -98,26 +95,35 @@ def from_geopandas(
     numbers = np.array(numbers)
 
     if abbrevs is not None:
-        if abbrevs == 'construct':
+        if abbrevs == "construct":
             abbrevs = _construct_abbrevs(geodataframe, names)
         else:
             abbrevs = _maybe_get_column(geodataframe, abbrevs)
     else:
-        raise ValueError('Please provide a string take can be taken from '
-                         'geodataframe.columns for `abbrevs`, '
-                         'found {}'.format(abbrevs))
+        raise ValueError(
+            "Please provide a string take can be taken from "
+            "geodataframe.columns for `abbrevs`, "
+            "found {}".format(abbrevs)
+        )
     if names is not None:
         names = _maybe_get_column(geodataframe, names)
     else:
-        raise ValueError('Please provide a string take can be taken from '
-                         'geodataframe.columns for `names`, found {}'.format(names))
+        raise ValueError(
+            "Please provide a string take can be taken from "
+            "geodataframe.columns for `names`, found {}".format(names)
+        )
 
-    outlines = _maybe_get_column(geodataframe, 'geometry')
+    outlines = _maybe_get_column(geodataframe, "geometry")
 
     # check duplicates
     for to_check in [abbrevs, names]:
         assert _check_duplicates(to_check)
 
-    return Regions(outlines,
-                   numbers=numbers, names=names,
-                   abbrevs=abbrevs, name=name, source=source)
+    return Regions(
+        outlines,
+        numbers=numbers,
+        names=names,
+        abbrevs=abbrevs,
+        name=name,
+        source=source,
+    )