+++ docs/whats_new.rst
@@ -31,6 +31,8 @@ Enhancements
 
 - New marine regions from natural earth added as :py:attr:`naturalearth.ocean_basins_50`
   (:pull:`63` by `Julius Busecke <https://github.com/jbusecke>`_)
+- Create regions from geopandas/shapefiles :py:attr:`from_geopandas`
+  (:pull:`101` by `Aaron Spring <https://github.com/aaronspring>`_)
 
 Bug Fixes
 ~~~~~~~~~